push

+++ front/src/store/modules/question.js
@@ -1,5 +1,8 @@
-import axios from "axios";
+import axios from "axios"
+import Swal from 'sweetalert2'
+
 const URL = 'http://localhost:3000/question'
+
 const header = { headers: {
   Authorization: 'Bearer ' + sessionStorage.getItem('token'),
 }}
@@ -10,28 +13,43 @@ const state = {
 
 const getters = {
   allQuestion: state => state.questions
-};
+}
 
 const actions = {
   async fetchQuestion({ commit }) {
-    const id = sessionStorage.getItem('quiz_id')
-    const response = await axios.get(`${URL}/${id}`, header);
-    commit('setQuestions', response.data);
+    const id = sessionStorage.getItem('quizId')
+    const response = await axios.get(`${URL}/${id}`, header)
+    if (response) {
+      commit('setQuestions', response.data)
+    }
   },
 
   async deleteQuestion({ commit }, id) {
-    await axios.delete(`${URL}/${id}`, header);
-    commit('deleteQuestion', id);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        axios.delete(`${URL}/${id}`, header)
+        commit('deleteQuestion', id)
+      }
+  })
+
   },
 
   async createQuestion({ commit }, body) {
-    await axios.post(URL, body, header);
-    commit('createQuestion', body);
+    await axios.post(URL, body, header)
+    commit('createQuestion', body)
   },
 
   async editQuestion({ commit }, body) {
-    await axios.put(`${URL}/${body.id}`, body,header);
-    commit('editQuestion', body);
+    await axios.put(`${URL}/${body.id}`, body,header)
+    commit('editQuestion', body)
   },
 }
 
@@ -40,9 +58,9 @@ const mutations = {
   deleteQuestion: (state, id) => (state.questions = state.questions.filter(question => question.id !== id)),
   createQuestion: (state, body) => (state.questions.push(body)),
   editQuestion: (state, updTodo) => {
-    const index = state.questions.findIndex(question => question.id === updTodo.id);
+    const index = state.questions.findIndex(question => question.id === updTodo.id)
     if (index !== -1) {
-      state.questions.splice(index, 1, updTodo);
+      state.questions.splice(index, 1, updTodo)
     }
   }
 }
@@ -52,4 +70,4 @@ export default {
   getters,
   actions,
   mutations,
-};
+}
