+++ front/src/store/modules/quiz.js
@@ -1,4 +1,5 @@
 import axios from "axios"
+import Swal from 'sweetalert2'
 
 const URL = "http://localhost:3000/quiz"
 
@@ -7,12 +8,16 @@ const header = { headers: {
 }}
 
 const state = {
-    quizzes: []
+    quizzes: [],
+    myQuizzes: [],
+    profile: ''
 }
 
 
 const getters = {
-    allQuiz: state => state.quizzes
+    allQuiz: state => state.quizzes,
+    myQuizzes: state => state.myQuizzes,
+    profile: state => state.profile
 }
 
 
@@ -22,37 +27,65 @@ const actions = {
         commit('setQuizzes', response.data)
     },
 
+    async fetchMyQuizzes({commit}) {
+        const response = await axios.get(`${URL}/myQuiz/${sessionStorage.getItem('userId')}`, header)
+        commit('setMyQuizzes', response.data)
+    },
+
     async editQuiz({commit}, body) {
         await axios.put(`${URL}/${body.id}`, body, header)
         commit('updateQuiz', body)
     },
 
     async createQuiz({commit}, body) {
-        console.log(body)
         await axios.post(`${URL}`, body, header)
-        // console.log(response.data)
         commit('pushQuiz', body)
     },
 
-    async deleteQuiz({commit}, id) {
-        await axios.delete(`${URL}/${id}`,header)
-        commit('removeQuiz', id)
+    deleteQuiz({commit}, id) {
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+                axios.delete(`${URL}/${id}`,header)
+                commit('removeQuiz', id)
+            }
+        })
     }
 }
 
     const mutations = {
     setQuizzes: (state, quiz) => (state.quizzes = quiz),
     
+    setMyQuizzes: (state, quiz) => (state.myQuizzes = quiz),
+
     updateQuiz: (state, body) => {
-        const index = state.quizzes.findIndex(quiz => quiz.id === body.id)
+        const index = state.myQuizzes.findIndex(quiz => quiz.id === body.id)
         if (index !== -1) {
-          state.quizzes.splice(index, 1, body)
+          state.myQuizzes.splice(index, 1, body)
         }
     },
 
-    pushQuiz: (state, body) => (state.quizzes.push(body)),
+    setProfile: (state, profile) => {state.profile = profile},
 
-    removeQuiz: (state, id) => (state.quizzes = state.quizzes.filter(quiz => quiz.id !== id)),
+    pushQuiz: (state, body) => {
+        console.log(body);
+        state.myQuizzes.unshift(body)
+        state.quizzes.unshift(body)
+
+    },
+
+    removeQuiz: (state, id) => {
+        state.myQuizzes = state.myQuizzes.filter(quiz => quiz.id !== id)
+        state.quizzes = state.quizzes.filter(quiz => quiz.id !== id)
+    },
+
 }
 
 export default {
